red.xml specific fixers make sense only when configuration file exists

+++ src/Eclipse-IDE/org.robotframework.ide.eclipse.main.plugin/src/org/robotframework/ide/eclipse/main/plugin/project/build/fix/RedConfigFileCompletionProposal.java
@@ -71,7 +71,7 @@ public abstract class RedConfigFileCompletionProposal implements ICompletionProp
             final Optional<RobotProjectConfig> openedConfig = robotProject.getOpenedProjectConfig();
             final RobotProjectConfig config = openedConfig.orElseGet(robotProject::getRobotProjectConfig);
 
-            if (apply(externalFile, config)) {
+            if (!config.isNullConfig() && apply(externalFile, config)) {
                 marker.delete();
 
                 if (!openedConfig.isPresent()) {
